pay attention to word order in fulltext search
